Add exact text matching before jaro_winkler
This attempts to speed up agent name parsing and matching by adding a new step to the hierarchy of agent lookups which attempts to find an exact-text match for agent names. This should be much faster than the fuzzy text implementation, but slower than the VIAF/LCNAF lookup process (which uses a cache).

If no exact match is found, we fall back to fuzzy matching. Ideally most agent names will already exist in the database and not resort to the fuzzy text stage